Added a bounty command

+++ src/main/java/com/thenexusreborn/survivalgames/game/Bounty.java
@@ -4,18 +4,18 @@ import java.util.*;
 
 public final class Bounty {
     private final UUID player;
-    private Map<Type, Double> amounts = new HashMap<>();
+    private Map<Type, Integer> amounts = new HashMap<>();
     
     public Bounty(UUID player) {
         this.player = player;
     }
     
     public double getAmount(Type type) {
-        return amounts.getOrDefault(type, 0.0);
+        return amounts.getOrDefault(type, 0);
     }
     
-    public void add(Type type, double amount) {
-        amounts.put(type, amounts.getOrDefault(type, 0.0) + amount);
+    public void add(Type type, int amount) {
+        amounts.put(type, amounts.getOrDefault(type, 0) + amount);
     }
     
     public UUID getPlayer() {
